meme card refactoring
