+++ src/app/(lobby)/memes/page.tsx
@@ -1,4 +1,4 @@
-import { MemeCardContainer } from "@/app/_components/meme-card-container"
+import { MemeCard } from "@/app/_components/meme-card"
 import { getServerAuthSession } from "@/server/auth"
 import { api } from "@/trpc/server"
 
@@ -9,7 +9,7 @@ const MemesPage = async () => {
         <div className="mx-auto flex w-full max-w-2xl flex-col space-y-8">
             {memes?.length ? (
                 memes.map((meme) => (
-                    <MemeCardContainer
+                    <MemeCard
                         meme={meme}
                         isAuthenticated={!!session?.user}
                         user={session?.user}
